Add polling logic to client
Client will now poll the socket connection to the server on every
iteration of the game loop to see if the server has sent an update.
If no update has been recived the client will render the most up to date
version of the game data which it has cached.